+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this extension will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.2.4] - 2025-03-05
+### Fixed
+- Css features fixes.
+
+### Added
+- Request history item click css.
+- Request history item now when clicked fills the fields of Request section.
+
 ## [0.2.3] - 2025-03-03
 ### Fixed
 - Css features fixes